Moves buttons from navigator and navigationBar into seperate components
Resolves #10354

Auditors: @cezaraugusto @bsclifton

Test Plan:
- check if reload button is working
- check if stop button is working
- check if forward/back buttons are working
- check if long press forward/back buttons are working
- check if bookmark button is working (add/remove)
