+++ app/renderer/components/navigation/navigator.js
@@ -3,8 +3,6 @@
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 const React = require('react')
-const electron = require('electron')
-const ipc = electron.ipcRenderer
 const Immutable = require('immutable')
 
 // Actions
@@ -14,11 +12,12 @@ const windowActions = require('../../../../js/actions/windowActions')
 // Components
 const ReduxComponent = require('../reduxComponent')
 const NavigationBar = require('./navigationBar')
-const LongPressButton = require('../common/longPressButton')
 const MenuBar = require('./menuBar')
-const WindowCaptionButtons = require('./windowCaptionButtons')
+const WindowCaptionButtons = require('./buttons/windowCaptionButtons')
 const Button = require('../common/button')
 const BrowserAction = require('./browserAction')
+const BackButton = require('./buttons/backButton')
+const ForwardButton = require('./buttons/forwardButton')
 
 // State
 const tabState = require('../../../common/state/tabState')
@@ -31,15 +30,12 @@ const windowState = require('../../../common/state/windowState')
 const {getCurrentWindowId, isMaximized, isFullScreen, isFocused} = require('../../currentWindow')
 const isWindows = require('../../../common/lib/platformUtil').isWindows()
 const {braveShieldsEnabled} = require('../../../common/state/shieldState')
-const eventUtil = require('../../../../js/lib/eventUtil')
-const {isNavigatableAboutPage, getBaseUrl} = require('./../../../../js/lib/appUrlUtil')
 const frameStateUtil = require('../../../../js/state/frameStateUtil')
 const siteSettings = require('../../../../js/state/siteSettings')
 const cx = require('../../../../js/lib/classSet')
 const {getSetting} = require('../../../../js/settings')
 
 // Constants
-const messages = require('../../../../js/constants/messages')
 const appConfig = require('../../../../js/constants/appConfig')
 const settings = require('../../../../js/constants/settings')
 
@@ -50,27 +46,10 @@ const globalStyles = require('../styles/global')
 class Navigator extends React.Component {
   constructor (props) {
     super(props)
-    this.onBack = this.onBack.bind(this)
-    this.onForward = this.onForward.bind(this)
-    this.onBackLongPress = this.onBackLongPress.bind(this)
-    this.onForwardLongPress = this.onForwardLongPress.bind(this)
     this.onDoubleClick = this.onDoubleClick.bind(this)
     this.onDragOver = this.onDragOver.bind(this)
     this.onDrop = this.onDrop.bind(this)
     this.onBraveMenu = this.onBraveMenu.bind(this)
-  }
-
-  onNav (e, navCheckProp, navType, navAction) {
-    if (e && eventUtil.isForSecondaryAction(e) && this.props.isNavigable) {
-      if (this.props[navCheckProp]) {
-        appActions.tabCloned(this.props.activeTabId, {
-          [navType]: true,
-          active: !!e.shiftKey
-        })
-      }
-    } else {
-      navAction.call(this, this.props.activeTabId)
-    }
   }
 
   get extensionButtons () {
@@ -79,30 +58,6 @@ class Navigator extends React.Component {
     buttons.push(<span className='buttonSeparator' />)
 
     return buttons
-  }
-
-  onBack (e) {
-    this.onNav(e, 'canGoBack', 'back', appActions.onGoBack)
-  }
-
-  onForward (e) {
-    this.onNav(e, 'canGoForward', 'forward', appActions.onGoForward)
-  }
-
-  onBackLongPress (target) {
-    const rect = target.parentNode.getBoundingClientRect()
-    appActions.onGoBackLong(this.props.activeTabId, {
-      left: rect.left,
-      bottom: rect.bottom
-    })
-  }
-
-  onForwardLongPress (target) {
-    const rect = target.parentNode.getBoundingClientRect()
-    appActions.onGoForwardLong(this.props.activeTabId, {
-      left: rect.left,
-      bottom: rect.bottom
-    })
   }
 
   onDragOver (e) {
@@ -137,23 +92,14 @@ class Navigator extends React.Component {
     if (!e.target.className.includes('navigatorWrapper')) {
       return
     }
-    return !isMaximized() ? windowActions.shouldMaximize(getCurrentWindowId()) : windowActions.shouldMinimize(getCurrentWindowId())
-  }
 
-  componentDidMount () {
-    ipc.on(messages.SHORTCUT_ACTIVE_FRAME_BACK, this.onBack)
-    ipc.on(messages.SHORTCUT_ACTIVE_FRAME_FORWARD, this.onForward)
-  }
-
-  componentWillUnmount () {
-    ipc.off(messages.SHORTCUT_ACTIVE_FRAME_BACK, this.onBack)
-    ipc.off(messages.SHORTCUT_ACTIVE_FRAME_FORWARD, this.onForward)
+    return !isMaximized()
+      ? windowActions.shouldMaximize(getCurrentWindowId())
+      : windowActions.shouldMinimize(getCurrentWindowId())
   }
 
   mergeProps (state, ownProps) {
     const currentWindow = state.get('currentWindow')
-    const swipeLeftPercent = state.get('swipeLeftPercent')
-    const swipeRightPercent = state.get('swipeRightPercent')
     const activeFrame = frameStateUtil.getActiveFrame(currentWindow) || Immutable.Map()
     const activeFrameKey = activeFrame.get('key')
     const activeTabId = activeFrame.get('tabId', tabState.TAB_ID_NONE)
@@ -173,8 +119,6 @@ class Navigator extends React.Component {
 
     const props = {}
     // used in renderer
-    props.canGoBack = activeTab.get('canGoBack') && !activeTabShowingMessageBox
-    props.canGoForward = activeTab.get('canGoForward') && !activeTabShowingMessageBox
     props.totalBlocks = activeFrame ? frameStateUtil.getTotalBlocks(activeFrame) : false
     props.shieldsDown = !braverySettings.shieldsUp
     props.shieldEnabled = braveShieldsEnabled(activeFrame)
@@ -193,21 +137,8 @@ class Navigator extends React.Component {
     props.isWideURLbarEnabled = getSetting(settings.WIDE_URL_BAR)
     props.showNavigationBar = activeFrameKey !== undefined &&
       state.get('siteSettings') !== undefined
-    props.swipeLeftPercent = swipeLeftPercent ? (swipeLeftPercent + 1) * 1.2 : 1
-    props.swipeRightPercent = swipeRightPercent ? (swipeRightPercent + 1) * 1.2 : 1
-    // 0.85 is the default button opacity in less/navigationBar.less
-    // Remove this magic number once we migrate to Aphrodite
-    props.swipeLeftOpacity = swipeLeftPercent ? 0.85 - (swipeLeftPercent > 0.65 ? 0.65 : swipeLeftPercent) : 0.85
-    props.swipeRightOpacity = swipeRightPercent ? 0.85 - (swipeRightPercent > 0.65 ? 0.65 : swipeRightPercent) : 0.85
-    if (swipeLeftPercent === 1) {
-      props.swipeLeftOpacity = 0.85
-    }
-    if (swipeRightPercent === 1) {
-      props.swipeRightOpacity = 0.85
-    }
 
     // used in other functions
-    props.isNavigable = activeFrame && isNavigatableAboutPage(getBaseUrl(activeFrame.get('location')))
     props.activeTabId = activeTabId
 
     return props
@@ -236,52 +167,8 @@ class Navigator extends React.Component {
             backforward: true,
             fullscreen: isFullScreen()
           })}>
-            <div data-test-id={
-              !this.props.canGoBack
-                ? 'navigationBackButtonDisabled'
-                : 'navigationBackButton'
-              }
-              className={cx({
-                navigationButtonContainer: true,
-                nav: true,
-                disabled: !this.props.canGoBack
-              })}
-              style={{
-                transform: this.props.canGoBack ? `scale(${this.props.swipeLeftPercent})` : `scale(1)`,
-                opacity: `${this.props.swipeLeftOpacity}`
-              }}>
-              <LongPressButton
-                testId={!this.props.canGoBack ? 'backButtonDisabled' : 'backButton'}
-                l10nId='backButton'
-                className='normalizeButton navigationButton backButton'
-                disabled={!this.props.canGoBack}
-                onClick={this.onBack}
-                onLongPress={this.onBackLongPress}
-              />
-            </div>
-            <div data-test-id={
-              !this.props.canGoForward
-                ? 'navigationForwardButtonDisabled'
-                : 'navigationForwardButton'
-              }
-              className={cx({
-                navigationButtonContainer: true,
-                nav: true,
-                disabled: !this.props.canGoForward
-              })}
-              style={{
-                transform: this.props.canGoForward ? `scale(${this.props.swipeRightPercent})` : `scale(1)`,
-                opacity: `${this.props.swipeRightOpacity}`
-              }}>
-              <LongPressButton
-                testId={!this.props.canGoForward ? 'forwardButtonDisabled' : 'forwardButton'}
-                l10nId='forwardButton'
-                className='normalizeButton navigationButton forwardButton'
-                disabled={!this.props.canGoForward}
-                onClick={this.onForward}
-                onLongPress={this.onForwardLongPress}
-              />
-            </div>
+            <BackButton />
+            <ForwardButton />
           </div>
           {
             this.props.showNavigationBar
